Refactor - move converter
